requirements.txt
streamlit
faiss-cpu
numpy
huggingface-hub
python-dotenv
